backend/agents/requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-ollama
perplexityai
python-dotenv
